Client+Shared: Move Language and Locale enums to shared package
These will be used in both the server and client so they need to be
moved to a place where they're available in both.
