+++ client/src/main/java/org/moparforia/client/Game.java
@@ -8,19 +8,14 @@ import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 import javax.swing.JFrame;
+import org.moparforia.shared.Locale;
 
 public class Game {
     private static final int WIDTH = 735;
     private static final int HEIGHT = 525;
 
     public Game(
-            JFrame frame,
-            String server,
-            int port,
-            Launcher.Locale locale,
-            String username,
-            boolean verbose,
-            boolean norandom) {
+            JFrame frame, String server, int port, Locale locale, String username, boolean verbose, boolean norandom) {
         Applet game = new AGolf();
 
         game.setStub(new Stub(server, locale, username, port, verbose, norandom));
@@ -38,8 +33,7 @@ public class Game {
         private final Map<String, String> params;
         private String server;
 
-        public Stub(
-                String server, Launcher.Locale locale, String username, int port, boolean verbose, boolean norandom) {
+        public Stub(String server, Locale locale, String username, int port, boolean verbose, boolean norandom) {
             if (server.indexOf(':') == -1) { // is ipv4
                 this.server = server;
             } else { // is ipv6
