+++ client/src/main/java/org/moparforia/client/Launcher.java
@@ -15,6 +15,7 @@ import javax.swing.JPasswordField;
 import javax.swing.JSpinner;
 import javax.swing.JTextField;
 import javax.swing.SpinnerNumberModel;
+import org.moparforia.shared.Locale;
 import org.moparforia.shared.ManifestVersionProvider;
 import picocli.CommandLine;
 
@@ -182,45 +183,5 @@ public class Launcher implements Callable<Integer> {
 
     public void setPort(int port) {
         this.port = port;
-    }
-
-    public enum Locale {
-        EN_US("en_US", Language.ENGLISH),
-        FI_FI("fi_FI", Language.FINNISH),
-        SV_SE("sv_SE", Language.SWEDISH);
-
-        private final String localeCode;
-        private final Language language;
-
-        Locale(String localeCode, Language language) {
-            this.localeCode = localeCode;
-            this.language = language;
-        }
-
-        public Language getLanguage() {
-            return this.language;
-        }
-
-        @Override
-        public String toString() {
-            return this.localeCode;
-        }
-    }
-
-    public enum Language {
-        ENGLISH("en"),
-        FINNISH("fi"),
-        SWEDISH("sv");
-
-        private final String languageCode;
-
-        Language(String languageCode) {
-            this.languageCode = languageCode;
-        }
-
-        @Override
-        public String toString() {
-            return this.languageCode;
-        }
     }
 }
