+++ shared/src/main/java/org/moparforia/shared/Language.java
@@ -1,0 +1,18 @@
+package org.moparforia.shared;
+
+public enum Language {
+    ENGLISH("en"),
+    FINNISH("fi"),
+    SWEDISH("sv");
+
+    private final String languageCode;
+
+    Language(String languageCode) {
+        this.languageCode = languageCode;
+    }
+
+    @Override
+    public String toString() {
+        return this.languageCode;
+    }
+}
